refactor: 우승자 결정 로직을 RacingGame Controller에서 RacingCars 도메인에서 처리하도록 수정
